Refactor metrics into a single package
